Разделяет код на модули (#8)
