+++ js/constants.js
@@ -1,0 +1,30 @@
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда выделяете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше!',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше!',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Иван',
+  'Степан',
+  'Мария',
+  'Дмитрий',
+  'Дарья',
+  'Ольга',
+];
+
+const DESCRIPTION = [
+  'В отпуске!',
+  'На природе',
+  'На работе!',
+  'С друзьями!',
+  'На прогулке!',
+  'Моя собака!',
+];
+
+const ALL_PHOTOS = 25;
+
+export { ALL_PHOTOS, DESCRIPTION, NAMES, MESSAGE };
